Transferred org class from main.py into seperate module for tidiness and future proofing
